+++ src/engraving/draw/painter.h
@@ -102,8 +102,8 @@ public:
     void drawPath(const QPainterPath& path);
     void strokePath(const QPainterPath& path, const QPen& pen);
 
-    void drawLines(const LineF* lines, int lineCount);
-    void drawLines(const PointF* pointPairs, int lineCount);
+    void drawLines(const LineF* lines, size_t lineCount);
+    void drawLines(const PointF* pointPairs, size_t lineCount);
     inline void drawLine(const LineF& line);
     inline void drawLine(const PointF& p1, const PointF& p2);
     inline void drawLines(const std::vector<LineF>& lines);
@@ -120,20 +120,20 @@ public:
     //! Fractions are also passed to this method, and, accordingly, the fractional part is discarded.
     inline void drawRect(int x1, int y1, int w, int h);
 
-    void drawRects(const RectF* rects, int rectCount);
+    void drawRects(const RectF* rects, size_t rectCount);
 
     void drawRoundedRect(const RectF& rect, qreal xRadius, qreal yRadius, Qt::SizeMode mode = Qt::AbsoluteSize);
 
     void drawEllipse(const RectF& rect);
     inline void drawEllipse(const PointF& center, qreal rx, qreal ry);
 
-    void drawPolyline(const PointF* points, int pointCount);
+    void drawPolyline(const PointF* points, size_t pointCount);
     inline void drawPolyline(const PolygonF& polyline);
 
-    void drawPolygon(const PointF* points, int pointCount, Qt::FillRule fillRule = Qt::OddEvenFill);
+    void drawPolygon(const PointF* points, size_t pointCount, Qt::FillRule fillRule = Qt::OddEvenFill);
     inline void drawPolygon(const PolygonF& polygon, Qt::FillRule fillRule = Qt::OddEvenFill);
 
-    void drawConvexPolygon(const PointF* points, int pointCount);
+    void drawConvexPolygon(const PointF* points, size_t pointCount);
     inline void drawConvexPolygon(const PolygonF& polygon);
 
     void drawArc(const RectF& rect, int a, int alen);
